Make http server pluggable.
